got rid of warnings

+++ src/pigsfli/src/pimc.cpp
@@ -305,7 +305,7 @@ int main(int argc, char** argv){
     // Adjacency matrix
     build_hypercube_adjacency_matrix(L,D,boundary_condition,adjacency_matrix);
     total_nn=0;
-    for (int i=0;i<adjacency_matrix[0].size();i++){total_nn+=1;}
+    for (size_t i=0;i<adjacency_matrix[0].size();i++){total_nn+=1;}
     
     // Replicated trackers
     for (int r=0;r<num_replicas;r++){
@@ -749,7 +749,7 @@ int main(int argc, char** argv){
         N_idx = N-N_min;
 
         // Fill out the histogram
-        for (int i=0;i<N_data.size();i++){
+        for (size_t i=0;i<N_data.size();i++){
             N_hist[N_data[i]-N_min]+=1;
             N_hist_sum+=1.0;
         }
@@ -757,7 +757,7 @@ int main(int argc, char** argv){
         // Build the normalized probability distribution P(N) & find its peak
         peak_idx=0;
         P_N_peak=P_N[peak_idx];
-        for (int i=0;i<P_N.size();i++){
+        for (size_t i=0;i<P_N.size();i++){
             P_N[i]=N_hist[i]/N_hist_sum;
             if (P_N[i]>P_N_peak){
                 peak_idx=i;
@@ -769,7 +769,7 @@ int main(int argc, char** argv){
         cout << "mu: " << mu;
         cout << " eta: " << eta << " Z-frac: " << Z_frac*100 << "%" << endl;
         cout << "N     P(N)"<<endl;
-        for (int i=0;i<N_bins.size();i++){
+        for (size_t i=0;i<N_bins.size();i++){
             cout << setw(6) << left << N_bins[i];
             for (int j=0;j<=static_cast<int>(100*P_N[i]);j++){
                 cout<<"*";
@@ -1574,7 +1574,7 @@ int main(int argc, char** argv){
                         if (writing_ctr==bin_size){
                             if (get_PN){
                             // Save total particle distribution
-                            for (int i=0; i<N_data.size(); i++){
+                            for (size_t i=0; i<N_data.size(); i++){
                                 PN_file<<fixed<<setprecision(17)<<N_data[i]<<" ";
                             }
                             PN_file<<endl;
